+++ Gateway/Request/TransactionReferenceDataBuilder.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (c) 2016-2019 Mastercard
+ * Copyright (c) 2016-2021 Mastercard
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,11 +17,10 @@
 
 namespace OnTap\MasterCard\Gateway\Request;
 
-use Magento\Payment\Gateway\Helper\SubjectReader;
 use Magento\Payment\Gateway\Request\BuilderInterface;
-use Magento\Quote\Model\Quote;
+use Magento\Payment\Gateway\Helper\SubjectReader;
 
-class OrderIdBuilder implements BuilderInterface
+class TransactionReferenceDataBuilder implements BuilderInterface
 {
     /**
      * @inheritDoc
@@ -29,30 +28,14 @@ class OrderIdBuilder implements BuilderInterface
     public function build(array $buildSubject)
     {
         $paymentDO = SubjectReader::readPayment($buildSubject);
-
-        /** @var Quote $quote */
-        $quote = $paymentDO->getPayment()->getQuote();
-
-        $this->refreshOrderIdReservation($quote);
+        $order = $paymentDO->getOrder();
+        $orderId = $order->getOrderIncrementId();
+        $txnId = uniqid(sprintf('%s-', $orderId));
 
         return [
-            'order' => [
-                'id' => $quote->getReservedOrderId()
+            'transaction' => [
+                'reference' => $txnId
             ]
         ];
-    }
-
-    /**
-     * Refresh Order Id Reservation
-     *
-     * If quote already used in payment gateway and failed
-     * then new session should contain another order id
-     *
-     * @param Quote $quote
-     */
-    private function refreshOrderIdReservation(Quote $quote)
-    {
-        $quote->setReservedOrderId('');
-        $quote->reserveOrderId();
     }
 }
